added form input as new component
