+++ src/App.jsx
@@ -1,26 +1,16 @@
 import './App.css'
+import ToDoList from './TodoList'
+import ToDoForm from './ToDoForm'
 
 function App() {
 
-  const todos = [
-    { id: 1, title: 'Learn React', completed: false },
-    { id: 2, title: 'Build a ToDo App', completed: false },
-    { id: 3, title: 'Share with friends', completed: false }
-  ]
+  
   return (
     <div>
       <h1>ToDo List</h1>
-      <ul>
-        {
-          todos.map((el) =>{
-            return(
-              <li key = {el.id}>
-                {el.title}
-              </li>
-            )
-          })
-        }
-      </ul>
+      <ToDoForm />
+      <ToDoList />
+
     </div>
   )
 }
